Move graphql resolvers and schemas to graphql folder - update docs to cover npm instructions - update scope of tests
