+++ pages/api/graphql.js
@@ -1,8 +1,7 @@
+import resolvers from '@/graphql/resolvers';
+import typeDefs from '@/graphql/schemas';
 import { ApolloServer } from '@apollo/server';
 import { startServerAndCreateNextHandler } from '@as-integrations/next';
-import resolvers from './resolvers';
-import typeDefs from './schemas';
-import React from 'react';
 
 const apolloServer = new ApolloServer({
   typeDefs,
